+++ docs/changelog.md
@@ -2,6 +2,7 @@
 
 ## 1.35.0 (26th November 2023)
 
+- Adds workspace scope validation - see https://docs.structurizr.com/workspaces for details.
 - structurizr/dsl: Updated to 1.34.0 - [changelog](https://github.com/structurizr/dsl/releases/tag/v1.34.0).
 - structurizr/export: Updated to 1.18.0 - [changelog](https://github.com/structurizr/export/releases/tag/v1.18.0).